// src/components/ActionsList/ActionsList.types.ts
// SPDX-FileCopyrightText: 2025 openstor contributors
// SPDX-FileCopyrightText: 2015-2025 MinIO, Inc.
// SPDX-License-Identifier: AGPL-3.0-or-later


import React from "react";
import { CSSObject } from "styled-components";

export interface ActionItem {
  action: () => void;
  label: string;
  disabled: boolean;
  icon: React.ReactNode;
  tooltip: string;
}

export interface ActionsListProps {
  sx?: CSSObject;
  items: ActionItem[];
  title: React.ReactNode;
}

export interface ActionsListPanelProps {
  sx?: CSSObject;
}

export interface ActionButtonProps
  extends React.ButtonHTMLAttributes<HTMLButtonElement> {
  label: string;
  icon: React.ReactNode;
}